+++ examples/bullet_prespawn/src/shared.rs
@@ -7,6 +7,7 @@ use leafwing_input_manager::prelude::ActionState;
 
 use lightyear::client::prediction::plugin::is_in_rollback;
 use lightyear::prelude::client::*;
+use lightyear::prelude::server::SyncTarget;
 use lightyear::prelude::TickManager;
 use lightyear::prelude::*;
 use lightyear::transport::io::IoDiagnosticsPlugin;
@@ -266,9 +267,8 @@ pub(crate) fn shoot_bullet(
                         //  The hash is computed automatically in PostUpdate from the entity's components + spawn tick
                         //  unless you set the hash manually before PostUpdate to a value of your choice
                         PreSpawnedPlayerObject::default(),
-                        ReplicateToServer {
-                            target: ReplicationTarget {
-                                target: NetworkTarget::All,
+                        Replicate {
+                            sync: SyncTarget {
                                 // the bullet is predicted for the client who shot it
                                 prediction: NetworkTarget::Single(id.0),
                                 // the bullet is interpolated for other clients
