+++ examples/client_replication/src/server.rs
@@ -121,11 +121,12 @@ pub(crate) fn replicate_players(
                 target: ReplicationTarget {
                     // we want to replicate back to the original client, since they are using a pre-spawned entity
                     target: NetworkTarget::All,
+                },
+                sync: SyncTarget {
                     // NOTE: even with a pre-spawned Predicted entity, we need to specify who will run prediction
-                    prediction: NetworkTarget::Only(vec![*client_id]),
+                    prediction: NetworkTarget::Single(*client_id),
                     // we want the other clients to apply interpolation for the player
                     interpolation: NetworkTarget::AllExceptSingle(*client_id),
-                    ..default()
                 },
                 ..default()
             };
@@ -155,6 +156,8 @@ pub(crate) fn replicate_cursors(
                 target: ReplicationTarget {
                     // do not replicate back to the client that owns the cursor!
                     target: NetworkTarget::AllExceptSingle(*client_id),
+                },
+                sync: SyncTarget {
                     // we want the other clients to apply interpolation for the cursor
                     interpolation: NetworkTarget::AllExceptSingle(*client_id),
                     ..default()
