+++ examples/leafwing_inputs/src/server.rs
@@ -109,16 +109,16 @@ pub(crate) fn replicate_players(
         // to other clients
         if let Some(mut e) = commands.get_entity(entity) {
             // we want to replicate back to the original client, since they are using a pre-predicted entity
-            let mut replication_target = ReplicationTarget::default();
+            let mut sync_target = SyncTarget::default();
 
             if global.predict_all {
-                replication_target.prediction = NetworkTarget::All;
+                sync_target.prediction = NetworkTarget::All;
             } else {
                 // we want the other clients to apply interpolation for the player
-                replication_target.interpolation = NetworkTarget::AllExceptSingle(client_id);
+                sync_target.interpolation = NetworkTarget::AllExceptSingle(client_id);
             }
             let replicate = Replicate {
-                target: replication_target,
+                sync: sync_target,
                 controlled_by: ControlledBy {
                     target: NetworkTarget::Single(client_id),
                 },
