+++ lightyear/src/client/components.rs
@@ -28,7 +28,7 @@ pub struct Confirmed {
 pub trait SyncComponent: Component + Clone + PartialEq + Message {}
 impl<T> SyncComponent for T where T: Component + Clone + PartialEq + Message {}
 
-/// Function that will interpolated between two values
+/// Function that will interpolate between two values
 pub trait LerpFn<C> {
     fn lerp(start: &C, other: &C, t: f32) -> C;
 }
