+++ lightyear/src/client/replication.rs
@@ -51,18 +51,22 @@ pub(crate) mod receive {
 
 pub(crate) mod send {
     use super::*;
-    use crate::prelude::server::ControlledBy;
+    use crate::prelude::server::{ControlledBy, ServerReplicationSet};
     use crate::prelude::{
-        ClientId, ComponentRegistry, ReplicateHierarchy, ReplicationGroup, ShouldBePredicted,
+        ClientId, ComponentRegistry, DisabledComponent, OverrideTargetComponent,
+        ReplicateHierarchy, ReplicateOnceComponent, ReplicationGroup, ShouldBePredicted,
         TargetEntity, VisibilityMode,
     };
     use crate::server::replication::send::SyncTarget;
     use crate::server::visibility::immediate::{ClientVisibility, ReplicateVisibility};
     use crate::shared::replication::components::{
-        Controlled, ReplicationTarget, ShouldBeInterpolated,
+        Controlled, DespawnTracker, Replicating, ReplicationTarget, ShouldBeInterpolated,
     };
     use crate::shared::replication::network_target::NetworkTarget;
-    use crate::shared::replication::ReplicationSend;
+    use crate::shared::replication::{systems, ReplicationSend};
+    use crate::shared::sets::ServerMarker;
+    use bevy::ecs::entity::Entities;
+    use bevy::ecs::system::SystemChangeTick;
 
     #[derive(Default)]
     pub struct ClientReplicationSendPlugin {
@@ -96,8 +100,29 @@ pub(crate) mod send {
                     send_entity_spawn
                         .in_set(InternalReplicationSet::<ClientMarker>::BufferEntityUpdates),
                 );
+            app.add_systems(
+                PostUpdate,
+                (
+                    // NOTE: we need to run `send_entity_despawn` once per frame (and not once per send_interval)
+                    //  because the RemovedComponents Events are present only for 1 frame and we might miss them if we don't run this every frame
+                    //  It is ok to run it every frame because it creates at most one message per despawn
+                    // NOTE: we make sure to update the replicate_cache before we make use of it in `send_entity_despawn`
+                    handle_replicating_remove
+                        .in_set(InternalReplicationSet::<ClientMarker>::BeforeBuffer),
+                    send_entity_despawn
+                        .in_set(InternalReplicationSet::<ClientMarker>::BufferDespawnsAndRemovals),
+                    handle_replicating_add
+                        .in_set(InternalReplicationSet::<ClientMarker>::AfterBuffer),
+                ),
+            );
         }
     }
+
+    /// Marker component that indicates that the entity should be replicated to the server
+    ///
+    /// If this component gets removed, we despawn the entity on the server.
+    #[derive(Component, Clone, Copy, Default, Debug, PartialEq, Reflect)]
+    pub struct ReplicationToServerTarget;
 
     /// Bundle that indicates how an entity should be replicated. Add this to an entity to start replicating
     /// it to the server.
@@ -118,8 +143,9 @@ pub(crate) mod send {
     /// - [`ReplicateHierarchy`] to specify how the hierarchy of the entity should be replicated
     #[derive(Bundle, Clone, Default, PartialEq, Debug, Reflect)]
     pub struct ReplicateToServer {
-        /// Should the entity be replicated to the client?
-        pub target: ReplicationTarget,
+        /// Marker indicating that the entity should be replicated to the server.
+        /// The replication stops if this component is removed
+        pub target: ReplicationToServerTarget,
         /// The replication group defines how entities are grouped (sent as a single message) for replication.
         ///
         /// After the entity is first replicated, the replication group of the entity should not be modified.
@@ -129,7 +155,62 @@ pub(crate) mod send {
         pub group: ReplicationGroup,
         /// How should the hierarchy of the entity (parents/children) be replicated?
         pub hierarchy: ReplicateHierarchy,
+        /// Marker indicating that we should send replication updates for that entity
+        pub replicating: Replicating,
     }
+
+    // TODO: replace this with observers
+    /// Metadata that holds Replicate-information from the previous send_interval's replication.
+    /// - when the entity gets despawned, we will use this to know how to replicate the despawn
+    #[derive(PartialEq, Debug)]
+    pub(crate) struct ReplicateCache {
+        pub(crate) replication_group: ReplicationGroup,
+    }
+
+    /// For every entity that removes their ReplicationTarget component but are not despawned, remove the component
+    /// from our replicate cache (so that the entity's despawns are no longer replicated)
+    pub(crate) fn handle_replicating_remove(
+        mut sender: ResMut<ConnectionManager>,
+        mut query: RemovedComponents<Replicating>,
+        entity_check: &Entities,
+    ) {
+        for entity in query.read() {
+            // only do this for entities that still exist
+            if entity_check.contains(entity) {
+                debug!("handling replicating component remove (delete from replicate cache)");
+                sender.replicate_component_cache.remove(&entity);
+            }
+        }
+    }
+
+    /// This system does all the additional bookkeeping required after [`Replicating`] has been added:
+    /// - adds DespawnTracker to each entity that was ever replicated, so that we can track when they are despawned
+    /// (we have a distinction between removing Replicating, which just stops replication; and despawning the entity)
+    /// - adds ReplicateCache for that entity so that when it's removed, we can know how to replicate the despawn
+    /// - adds the ReplicateVisibility component if needed
+    pub(crate) fn handle_replicating_add(
+        mut sender: ResMut<ConnectionManager>,
+        mut commands: Commands,
+        // We use `(With<Replicating>, Without<DespawnTracker>)` as an optimization instead of
+        // `Added<Replicating>`
+        query: Query<(Entity, &ReplicationGroup), (With<Replicating>, Without<DespawnTracker>)>,
+    ) {
+        for (entity, group) in query.iter() {
+            debug!("Replicate component was added for entity {entity:?}");
+            commands.entity(entity).insert(DespawnTracker);
+            sender.replicate_component_cache.insert(
+                entity,
+                ReplicateCache {
+                    replication_group: *group,
+                },
+            );
+        }
+    }
+
+    /// In host-server mode, the pre-predicted entities are spawned on the client.
+    /// The networking systems don't run, so we need to send events to the server
+    /// to make it seem like they received a replicated entity.
+    pub(crate) fn handle_pre_predicted_entity_host_server() {}
 
     /// Send entity spawn replication messages to server when the ReplicationTarget component is added
     /// Also handles:
@@ -137,48 +218,282 @@ pub(crate) mod send {
     /// - setting the priority
     pub(crate) fn send_entity_spawn(
         query: Query<
-            (
-                Entity,
-                Ref<ReplicationTarget>,
-                &ReplicationGroup,
-                Option<&TargetEntity>,
-            ),
-            Changed<ReplicationTarget>,
+            (Entity, &ReplicationGroup, Option<&TargetEntity>),
+            (With<Replicating>, Added<ReplicationToServerTarget>),
         >,
         mut sender: ResMut<ConnectionManager>,
     ) {
-        query
-            .iter()
-            .for_each(|(entity, replication_target, group, target_entity)| {
-                let mut target = replication_target.target.clone();
-                if !replication_target.is_added() {
-                    if let Some(cached_replicate) = sender.replicate_component_cache.get(&entity) {
-                        // do not re-send a spawn message to the server if we already have sent one
-                        target.exclude(&cached_replicate.replication_target)
-                    }
-                }
-                if target.is_empty() {
-                    return;
-                }
-                trace!(?entity, "Prepare entity spawn to server");
-                let group_id = group.group_id(Some(entity));
-                if let Some(TargetEntity::Preexisting(remote_entity)) = target_entity {
-                    sender.replication_sender.prepare_entity_spawn_reuse(
-                        entity,
-                        group_id,
-                        *remote_entity,
-                    );
-                } else {
-                    sender
-                        .replication_sender
-                        .prepare_entity_spawn(entity, group_id);
-                }
-                // TODO: should the priority be a component on the entity? but it should be shared between a group
-                //  should a GroupChannel be a separate entity?
-                // also set the priority for the group when we spawn it
+        query.iter().for_each(|(entity, group, target_entity)| {
+            trace!(?entity, "Prepare entity spawn to server");
+            let group_id = group.group_id(Some(entity));
+            if let Some(TargetEntity::Preexisting(remote_entity)) = target_entity {
+                sender.replication_sender.prepare_entity_spawn_reuse(
+                    entity,
+                    group_id,
+                    *remote_entity,
+                );
+            } else {
                 sender
                     .replication_sender
-                    .update_base_priority(group_id, group.priority());
-            });
+                    .prepare_entity_spawn(entity, group_id);
+            }
+            // TODO: should the priority be a component on the entity? but it should be shared between a group
+            //  should a GroupChannel be a separate entity?
+            // also set the priority for the group when we spawn it
+            sender
+                .replication_sender
+                .update_base_priority(group_id, group.priority());
+        });
+    }
+
+    /// Send entity despawn if:
+    /// - an entity that had a DespawnTracker was despawned
+    /// - an entity with Replicating had the ReplicationToServerTarget removed
+    pub(crate) fn send_entity_despawn(
+        query: Query<
+            (Entity, &ReplicationGroup),
+            (With<Replicating>, Without<ReplicationToServerTarget>),
+        >,
+        // TODO: ideally we want to send despawns for entities that still had REPLICATE at the time of despawn
+        //  not just entities that had despawn tracker once
+        mut despawn_removed: RemovedComponents<DespawnTracker>,
+        mut sender: ResMut<ConnectionManager>,
+    ) {
+        // Send entity-despawn for entities that have Replicating
+        // but where the `ReplicationToServerTarget` component has been removed
+        query.iter().for_each(|(entity, group)| {
+            trace!(
+                ?entity,
+                "send entity despawn because ReplicationToServerTarget was removed"
+            );
+            sender
+                .replication_sender
+                .prepare_entity_despawn(entity, group.group_id(Some(entity)));
+        });
+
+        // Despawn entities when the entity gets despawned on local world
+        for entity in despawn_removed.read() {
+            // only replicate the despawn if the entity still had a Replicating component
+            if let Some(replicate_cache) = sender.replicate_component_cache.remove(&entity) {
+                trace!(
+                    ?entity,
+                    "send entity despawn because DespawnTracker component was removed"
+                );
+                sender.replication_sender.prepare_entity_despawn(
+                    entity,
+                    replicate_cache.replication_group.group_id(Some(entity)),
+                );
+            }
+        }
+    }
+
+    /// This system sends updates for all components that were added or changed
+    /// Sends both ComponentInsert for newly added components and ComponentUpdates otherwise
+    ///
+    /// Updates are sent only for any components that were changed since the most recent of:
+    /// - last time we sent an update for that group which got acked.
+    ///
+    /// NOTE: cannot use ConnectEvents because they are reset every frame
+    pub(crate) fn send_component_update<C: Component>(
+        registry: Res<ComponentRegistry>,
+        query: Query<
+            (
+                Entity,
+                Ref<C>,
+                Ref<ReplicationToServerTarget>,
+                &ReplicationGroup,
+                Has<DisabledComponent<C>>,
+                Has<ReplicateOnceComponent<C>>,
+            ),
+            With<Replicating>,
+        >,
+        system_bevy_ticks: SystemChangeTick,
+        mut sender: ResMut<ConnectionManager>,
+    ) {
+        let kind = registry.net_id::<C>();
+        query.iter().for_each(
+            |(entity, component, target, group, disabled, replicate_once)| {
+                // do not replicate components that are disabled
+                if disabled {
+                    return;
+                }
+                let (mut insert, mut update) = (false, false);
+
+                // send a component_insert for components that were newly added
+                // or if we start replicating the entity
+                if component.is_added() || target.is_added() {
+                    trace!("component is added or replication_target is added");
+                    insert = true;
+                } else {
+                    // do not send updates for these components, only inserts/removes
+                    if replicate_once {
+                        trace!(?entity,
+                        "not replicating updates for {:?} because it is marked as replicate_once",
+                        kind
+                    );
+                        return;
+                    }
+                    // otherwise send an update for all components that changed since the
+                    // last update we have ack-ed
+                    update = true;
+                }
+                if insert || update {
+                    // serialize component
+                    let writer = sender.writer();
+                    let raw_data = registry
+                        .serialize(component.as_ref(), writer)
+                        .expect("Could not serialize component");
+
+                    if insert {
+                        let group_id = group.group_id(Some(entity));
+                        sender
+                            .replication_sender
+                            .prepare_component_insert(entity, group_id, kind, raw_data);
+                    } else {
+                        let group_id = group.group_id(Some(entity));
+                        // TODO: should we have additional state tracking so that we know we are in the process of sending this entity to clients?
+                        let collect_changes_since_this_tick = sender
+                            .replication_sender
+                            .group_channels
+                            .entry(group_id)
+                            .or_default()
+                            .collect_changes_since_this_tick;
+
+                        // send the update for all changes newer than the last ack bevy tick for the group
+                        if collect_changes_since_this_tick.map_or(true, |c| {
+                            component
+                                .last_changed()
+                                .is_newer_than(c, system_bevy_ticks.this_run())
+                        }) {
+                            trace!(
+                                change_tick = ?component.last_changed(),
+                                ?collect_changes_since_this_tick,
+                                current_tick = ?system_bevy_ticks.this_run(),
+                                "prepare entity update changed check"
+                            );
+                            // trace!(
+                            //     ?entity,
+                            //     component = ?kind,
+                            //     tick = ?self.tick_manager.tick(),
+                            //     "Updating single component"
+                            // );
+                            sender
+                                .replication_sender
+                                .prepare_entity_update(entity, group_id, kind, raw_data);
+                        }
+                    }
+                }
+            },
+        );
+    }
+
+    /// Send component remove
+    pub(crate) fn send_component_removed<C: Component>(
+        registry: Res<ComponentRegistry>,
+        // only remove the component for entities that are being actively replicated
+        query: Query<
+            (&ReplicationGroup, Has<DisabledComponent<C>>),
+            (With<Replicating>, With<ReplicationToServerTarget>),
+        >,
+        mut removed: RemovedComponents<C>,
+        mut sender: ResMut<ConnectionManager>,
+    ) {
+        let kind = registry.net_id::<C>();
+        removed.read().for_each(|entity| {
+            if let Ok((group, disabled)) = query.get(entity) {
+                // do not replicate components that are disabled
+                if disabled {
+                    return;
+                }
+                let group_id = group.group_id(Some(entity));
+                trace!(?entity, kind = ?std::any::type_name::<C>(), "Sending RemoveComponent");
+                sender
+                    .replication_sender
+                    .prepare_component_remove(entity, group_id, kind);
+            }
+        })
+    }
+
+    pub(crate) fn register_replicate_component_send<C: Component>(app: &mut App) {
+        app.add_systems(
+            PostUpdate,
+            (
+                // NOTE: we need to run `send_component_removed` once per frame (and not once per send_interval)
+                //  because the RemovedComponents Events are present only for 1 frame and we might miss them if we don't run this every frame
+                //  It is ok to run it every frame because it creates at most one message per despawn
+                crate::server::replication::send::send_component_removed::<C>
+                    .in_set(InternalReplicationSet::<ClientMarker>::BufferDespawnsAndRemovals),
+                // NOTE: we run this system once every `send_interval` because we don't want to send too many Update messages
+                //  and use up all the bandwidth
+                crate::server::replication::send::send_component_update::<C>
+                    .in_set(InternalReplicationSet::<ClientMarker>::BufferComponentUpdates),
+            ),
+        );
+    }
+
+    #[cfg(test)]
+    mod tests {
+        use crate::prelude::{client, server, ClientId};
+        use crate::tests::stepper::{BevyStepper, Step, TEST_CLIENT_ID};
+
+        #[test]
+        fn test_entity_spawn_client_to_server() {
+            let mut stepper = BevyStepper::default();
+
+            // spawn an entity on server with visibility::All
+            let client_entity = stepper.client_app.world.spawn_empty().id();
+            stepper.frame_step();
+            stepper.frame_step();
+
+            // add replicate
+            stepper
+                .client_app
+                .world
+                .entity_mut(client_entity)
+                .insert(client::ReplicateToServer::default());
+            // TODO: we need to run a couple frames because the server doesn't read the client's updates
+            //  because they are from the future
+            stepper.frame_step();
+            stepper.frame_step();
+            stepper.frame_step();
+
+            // check that the entity was spawned
+            stepper
+                .server_app
+                .world
+                .resource::<server::ConnectionManager>()
+                .connection(ClientId::Netcode(TEST_CLIENT_ID))
+                .expect("client connection missing")
+                .replication_receiver
+                .remote_entity_map
+                .get_local(client_entity)
+                .expect("entity was not replicated to server");
+        }
+    }
+}
+
+pub(crate) mod commands {
+    use crate::client::connection::ConnectionManager;
+    use crate::shared::replication::ReplicationSend;
+    use bevy::ecs::system::{Command, EntityCommands};
+    use bevy::prelude::{Entity, World};
+
+    fn despawn_without_replication(entity: Entity, world: &mut World) {
+        let mut sender = world.resource_mut::<ConnectionManager>();
+        // remove the entity from the cache of entities that are being replicated
+        // so that if it gets despawned, the despawn won't be replicated
+        sender.replicate_component_cache.remove(&entity);
+        world.despawn(entity);
+    }
+
+    pub trait DespawnReplicationCommandExt {
+        /// Despawn the entity and makes sure that the despawn won't be replicated.
+        fn despawn_without_replication(&mut self);
+    }
+
+    impl DespawnReplicationCommandExt for EntityCommands<'_> {
+        fn despawn_without_replication(&mut self) {
+            self.add(despawn_without_replication);
+        }
     }
 }
